Formatted Category CRUD pages

+++ MVCReviewProject/Controllers/CategoriesController.cs
@@ -109,9 +109,14 @@ namespace MVCReviewProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Category category = db.Categories.Find(id);
-            db.Categories.Remove(category);
-            db.SaveChanges();
+            // This prevents the main 8 islands from being deleted
+            if (id > 8)
+            {
+                Category category = db.Categories.Find(id);
+                db.Categories.Remove(category);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
             return RedirectToAction("Index");
         }
